feat(social-graph): implement TASK-3.16 custom hooks and profile updates

+++ backend/src/main/java/com/instagram/adapter/in/web/dto/response/UserProfileResponse.java
@@ -7,11 +7,11 @@ import io.swagger.v3.oas.annotations.media.Schema;
 public record UserProfileResponse(
         @Schema(description = "Basic user details") UserResponse user,
 
-        @Schema(description = "Number of posts created by the user", example = "42") int postCount,
+        @Schema(description = "Number of posts created by the user", example = "42") long postCount,
 
-        @Schema(description = "Number of followers", example = "1000") int followerCount,
+        @Schema(description = "Number of followers", example = "1000") long followerCount,
 
-        @Schema(description = "Number of accounts the user is following", example = "500") int followingCount,
+        @Schema(description = "Number of accounts the user is following", example = "500") long followingCount,
 
         @Schema(description = "Whether the current authenticated user is following this user", example = "true") boolean isFollowing) {
     public static UserProfileResponse from(UserProfile profile) {
