+++ backend/src/main/java/com/instagram/domain/model/UserStats.java
@@ -2,7 +2,8 @@ package com.instagram.domain.model;
 
 import java.util.UUID;
 
-public record UserStats(UUID userId, int postCount, int followerCount, int followingCount) {
+public record UserStats(UUID userId, long postCount, long followerCount, long followingCount) {
+
     public static UserStats zero(UUID userId) {
         return new UserStats(userId, 0, 0, 0);
     }
